+++ web/src/main/java/com/navercorp/pinpoint/web/calltree/span/SpanEventAlign.java
@@ -166,7 +166,7 @@ public class SpanEventAlign implements Align {
     }
 
     @Override
-    public Short getApplicationServiceType() {
+    public int getApplicationServiceType() {
         return spanBo.getApplicationServiceType();
     }
 
@@ -176,7 +176,7 @@ public class SpanEventAlign implements Align {
     }
 
     @Override
-    public short getServiceType() {
+    public int getServiceType() {
         return spanEventBo.getServiceType();
     }
