+++ web/src/main/java/com/navercorp/pinpoint/web/service/component/DefaultLegacyAgentCompatibility.java
@@ -1,7 +1,5 @@
 package com.navercorp.pinpoint.web.service.component;
 
-import com.navercorp.pinpoint.common.server.util.time.Range;
-
 import java.util.List;
 
 public class DefaultLegacyAgentCompatibility implements LegacyAgentCompatibility {
@@ -9,16 +7,16 @@ public class DefaultLegacyAgentCompatibility implements LegacyAgentCompatibility
 
     // legacy node agent support
     private final LegacyAgent[] legacyAgents = {
-            new LegacyAgent(List.of((short)1400, (short)1401), "0.8.0"),
+            new LegacyAgent(List.of(1400, 1401), "0.8.0"),
     };
 
     @Override
-    public boolean isLegacyAgent(short serviceType) {
+    public boolean isLegacyAgent(int serviceType) {
         return isLegacyAgent(serviceType, null);
     }
 
     @Override
-    public boolean isLegacyAgent(short serviceType, String version) {
+    public boolean isLegacyAgent(int serviceType, String version) {
         for (LegacyAgent legacyAgent : legacyAgents) {
             if (!legacyAgent.isLegacyType(serviceType)) {
                 return false;
